pyproject.toml
[project]
name = "ends_with"
version = "0.0.0"
description = "Validates that a list or string ends with a given value."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8"
dependencies = [
    "rstr",
    "guardrails-ai>=0.3.2",
    "pydantic>=2.4.2"
]